fix(DataTable): resolve conflicts
